createLinkedList returns a linked list object
createLinkedList accepts a set of ordered values and returns a
linked list object.

The head property of the linked list contains the first node in the
list.

The nodes are linked such that the first node contains the first value
and links to the node containing the following value. The node for the
last value contains a null link.