addTwo() returns a linked list for the sum of two linked lists
addTwo accepts the first node of two linked lists for numbers where
each node is a digit for a base 10 number. The first node for the
sum of the two numbers is returned.

This is the first solution developed. This solution does not
edit the constructor for the ListNode in any way and does not store a
value greater than 9 as the value for a node at any point during
runtime.